[Update] Fix a “Trying to get property 'id' of non-object” warning

+++ Joomla4/lscache_plugin/lscache.php
@@ -199,14 +199,14 @@ class plgSystemLSCache extends CMSPlugin {
      //if post back, purge current page, disabled in case purge search post back
         if ($this->pageCachable && ($app->input->getMethod() != 'GET')) {
             $this->pageCachable = false;
-            if ($this->menuItem && ($this->settings->get('purgePostBack', 0) == 1) ) {
+            if ($this->menuItem && isset($this->menuItem->id) && ($this->settings->get('purgePostBack', 0) == 1) ) {
                 $purgeTags = "com_menus:" . $this->menuItem->id;
                 $this->lscInstance->purgePublic($purgeTags);
                 $session->set('lastPostBack', $this->menuItem->id);
                 $this->log();
             }
         } else {
-            if($this->menuItem && ($session->get('lastPostBack')==$this->menuItem->id)){
+            if($this->menuItem && isset($this->menuItem->id) && ($session->get('lastPostBack')==$this->menuItem->id)){
                 $this->pageCachable = false;
                 $session->clear('lastPostBack');
             }
